ReorganizeAndRunSnake
- Files reorganized -Snake game running
