--- .gitignore
@@ -1,4 +1,0 @@
-.vscode/.browse.VC.db
-.vscode/.browse.VC.db-wal
-lcd_snake/.vscode/.browse.VC.db
-lcd_snake/.vscode/.browse.VC.db-wal
