+++ arcade/game.h
@@ -1,0 +1,8 @@
+
+class Game
+{
+public:
+    virtual void updateLCD(LCD *lcd) { }
+    virtual void changeDirection(byte direction) { }
+    virtual void checkCatchSeed() { }
+};
