+++ arcade/space_invaders.h
@@ -1,0 +1,74 @@
+#define MAX_ENEMIES     5
+
+class Enemy
+{
+private:
+    byte x;
+    byte y;
+public:
+    Enemy() { }
+    Enemy(byte x, byte y)
+    {
+        this->x = x;
+        this->y = y;
+    }
+    void move()
+    {
+
+    }
+};
+class PlayerShip
+{
+private:
+    byte x;
+    byte y;
+public:
+    PlayerShip() { }
+    PlayerShip(byte x, byte y)
+    {
+        this->x = x;
+        this->y = y;
+    }
+};
+class SpaceInvaders : public Game
+{
+private:
+    PlayerShip player_ship;
+    Enemy enemies[MAX_ENEMIES];
+public:
+    SpaceInvaders(byte x, byte y)
+    {
+        player_ship = PlayerShip(x, y);
+        //
+        randomSeed(millis());
+    }
+    void updateLCD(LCD *lcd)
+    {
+
+    }
+    void moveEnemies()
+    {
+        byte e;
+        for(e = 0; e < MAX_ENEMIES; e++)
+        {
+            enemies[e].move();
+        }
+    }
+};
+
+const byte ship_image[8] = {
+	0x00, 0xF0, 0xFC, 0x3F, 0x3F, 0xFC, 0xF0, 0x00
+};
+
+const byte enemy_type1_image[8] = {
+	0xF1, 0xBD, 0xB6, 0x3C, 0x3C, 0xB6, 0xBD, 0xF1
+};
+
+const byte spaceinvaders_load_image[6 * 84] = {
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0xF0, 0xF8, 0xF8, 0xF8, 0xF8, 0xF8, 0xF8, 0xF0, 0xF0, 0xF0, 0xE0, 0xE0, 0x00, 0x00, 0x00, 0xF8, 0xF8, 0xF8, 0xF8, 0xF8, 0x78, 0x78, 0x78, 0x78, 0xF8, 0xF8, 0xF0, 0xE0, 0xE0, 0x00, 0x80, 0xE0, 0xE0, 0xFC, 0xFC, 0xFC, 0xFC, 0xFC, 0xE0, 0xE0, 0x80, 0x00, 0x00, 0x00, 0x00, 0xE0, 0xF0, 0xF8, 0xF8, 0xFC, 0x7C, 0x7C, 0x7C, 0xFC, 0xFC, 0xF8, 0xF8, 0xF0, 0x00, 0x80, 0x80, 0xFC, 0xFC, 0xFC, 0x7C, 0x7C, 0x7C, 0x7C, 0x7C, 0x7C, 0x7C, 0x7C, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x03, 0x0F, 0x1F, 0x1E, 0x9E, 0xBE, 0xBC, 0x3B, 0x3B, 0x73, 0xF3, 0xF3, 0xE0, 0x00, 0x00, 0x0F, 0xFF, 0xFF, 0xFF, 0xFF, 0xFC, 0x3C, 0x3C, 0x3C, 0x3F, 0x3F, 0x0F, 0x07, 0x07, 0x00, 0x8F, 0xFF, 0xFF, 0xFB, 0xFF, 0x3F, 0xFB, 0xFB, 0xFF, 0xFF, 0xCF, 0x00, 0x00, 0x00, 0xF8, 0xFF, 0xFF, 0xFF, 0xFF, 0x07, 0x80, 0x80, 0x80, 0x83, 0x83, 0x03, 0x03, 0x03, 0xF8, 0xFF, 0xFF, 0x1F, 0x1F, 0x1F, 0x1C, 0x1C, 0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0xC0, 0xC0, 0x03, 0xC7, 0xC7, 0xCE, 0xCE, 0x0E, 0x0F, 0x0F, 0xC7, 0xC0, 0xC0, 0x00, 0xCF, 0xCF, 0xCF, 0xCF, 0x0F, 0x00, 0xC0, 0xC0, 0xC0, 0xC0, 0x00, 0x80, 0xC2, 0xC2, 0xC3, 0xCF, 0xCF, 0x8E, 0x8E, 0x00, 0xCE, 0xCE, 0xCF, 0xCF, 0xC3, 0xC2, 0xC0, 0x80, 0x03, 0xC3, 0xCF, 0xCF, 0xCF, 0x4E, 0x4F, 0x4F, 0x43, 0x43, 0x83, 0x80, 0xCF, 0xCF, 0xCF, 0xCF, 0xCF, 0x8E, 0x0E, 0x0E, 0x8E, 0xCE, 0xC0, 0x40, 0xC0, 0xC0, 0x80, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x7F, 0x7F, 0x7F, 0x00, 0x7F, 0x7F, 0x7F, 0x7F, 0x07, 0x0E, 0x0E, 0x7F, 0x7F, 0x7F, 0x00, 0x03, 0x1F, 0x1F, 0x7F, 0x7C, 0x7C, 0x7F, 0x1F, 0x03, 0x03, 0x00, 0x7F, 0x7F, 0x7F, 0x0D, 0x7F, 0x7F, 0x7F, 0x7F, 0x00, 0x7F, 0x7F, 0x7F, 0x7F, 0x60, 0x7F, 0x7F, 0x1F, 0x00, 0x7F, 0x7F, 0x7F, 0x7F, 0x44, 0x44, 0x44, 0x44, 0x44, 0x7F, 0x7F, 0x7F, 0x7F, 0x0D, 0x7F, 0x7F, 0x7B, 0x00, 0x00, 0x13, 0x77, 0x77, 0x64, 0x7C, 0x7C, 0x18, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00,
+	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
