--- lcd_snake/game.h
@@ -1,6 +1,0 @@
-
-class Game
-{
-public:
-    virtual void updateLCD(LCD *lcd) { }
-};
