--- lcd_snake/snake.h
@@ -1,271 +1,0 @@
-#define DIRECTION_NORTH     1
-#define DIRECTION_SOUTH     2
-#define DIRECTION_EAST      3
-#define DIRECTION_WEST      4
-//
-#define INITIAL_SNAKE_BODY_SIZE     8
-#define INITIAL_SNAKE_X             5
-#define INITIAL_SNAKE_Y             10
-//
-struct body_move_
-{
-    byte order;
-    byte direction;
-    struct body_move_ *next;
-};
-typedef struct body_move_ BodyMove;
-
-struct body_snake_
-{
-    byte x;
-    byte y;
-    byte direction;
-    struct body_snake_ *next;
-    BodyMove *next_moves;
-};
-typedef struct body_snake_ BodySnake;
-
-class Snake
-{
-private:
-    BodySnake *head;
-    BodySnake *tail;
-    BodySnake *body_snake;
-public:
-    Snake() { }
-    Snake(byte x, byte y)
-    {
-        body_snake = (BodySnake*)malloc(sizeof(BodySnake));
-        body_snake->x = x;
-        body_snake->y = y;
-        body_snake->direction = DIRECTION_SOUTH;
-        body_snake->next = NULL;
-        body_snake->next_moves = NULL;
-
-        head = body_snake;
-        tail = body_snake;
-    }
-    void growth()
-    {
-        BodySnake *new_body_part;
-        BodyMove *new_body_move;
-        BodyMove *tail_moves;
-        byte x, y;
-        x = y = 0;
-
-        switch(tail->direction)
-        {
-            case DIRECTION_NORTH:
-                y ++;
-                break;
-            case DIRECTION_SOUTH:
-                y --;
-                break;
-            case DIRECTION_EAST:
-                x --;
-                break;
-            case DIRECTION_WEST:
-                x ++;
-                break;
-        }
-
-        new_body_part = (BodySnake*)malloc(sizeof(BodySnake));
-        new_body_part->x = tail->x + x;
-        new_body_part->y = tail->y + y;
-        new_body_part->direction = tail->direction;
-        new_body_part->next = NULL;
-        new_body_part->next_moves = NULL;
-
-        tail_moves = tail->next_moves;
-        while(tail_moves != NULL)
-        {
-            new_body_move = (BodyMove*)malloc(sizeof(BodyMove));
-            new_body_move->order = tail_moves->order + 1;
-            new_body_move->direction = tail_moves->direction;
-
-            new_body_move->next = new_body_part->next_moves;
-            new_body_part->next_moves = new_body_move;
-
-            tail_moves = tail_moves->next;
-        }
-
-        tail->next = new_body_part;
-        tail = new_body_part;
-    }
-    BodySnake* getHead()
-    {
-        return head;
-    }
-    BodySnake* getTail()
-    {
-        return tail;
-    }
-    void setDirection(byte direction)
-    {
-        BodyMove *new_next_move;
-        BodySnake *body_part;
-        byte order = 1;
-
-        body_part = head;
-        while(body_part != NULL)
-        {
-            new_next_move = (BodyMove*)malloc(sizeof(BodyMove));
-            new_next_move->order = order;
-            new_next_move->direction = direction;
-
-            new_next_move->next = body_part->next_moves;
-            body_part->next_moves = new_next_move;
-
-            order ++;
-            body_part = body_part->next;
-        }
-    }
-    void updateDirections()
-    {
-        BodyMove *move, *prev_move;
-        BodySnake *body_part;
-
-        body_part = head;
-        while(body_part != NULL)
-        {
-            move = body_part->next_moves;
-            prev_move = move;
-            while(move != NULL)
-            {
-                move->order --;
-                if(move->order == 0)
-                {
-                    int direction = move->direction;
-                    body_part->direction = direction;
-
-                    if(prev_move == body_part->next_moves)
-                        body_part->next_moves = NULL;
-                    else
-                        prev_move->next = NULL;
-
-                    free(move);
-                    break;
-                }
-                else
-                {
-                    prev_move = move;
-                    move = move->next;
-                }
-            }
-
-            body_part = body_part->next;
-        }
-    }
-    void move()
-    {
-        BodySnake *snake_part = body_snake;
-        while(snake_part != NULL)
-        {
-            switch(snake_part->direction)
-            {
-                case DIRECTION_NORTH:
-                    if(snake_part->y == 0)
-                        snake_part->y = LCD_Y;
-                    snake_part->y --;
-                    break;
-                case DIRECTION_SOUTH:
-                    snake_part->y ++;
-                    if(snake_part->y == LCD_Y)
-                        snake_part->y = 0;
-                    break;
-                case DIRECTION_EAST:
-                    snake_part->x ++;
-                    if(snake_part->x == LCD_X)
-                        snake_part->x = 0;
-                    break;
-                case DIRECTION_WEST:
-                    if(snake_part->x == 0)
-                        snake_part->x = LCD_X;
-                    snake_part->x --;
-                    break;
-            }
-            //
-            snake_part = snake_part->next;
-                
-        }
-    }
-};
-
-class SnakeGame : public Game
-{
-private:
-    Snake snake;
-    byte snake_food[2];
-public:
-    SnakeGame(byte x, byte y, byte length) 
-    {
-        byte i = 0;
-        snake = Snake(x, y);
-        while(i++ < length)
-            snake.growth();
-
-        randomSeed(millis());
-        generateSeed();
-    }
-    void updateLCD(LCD *lcd)
-    {
-        BodySnake *body_snake = snake.getHead();
-        
-        lcd->clear();
-        while(body_snake != NULL)
-        {
-            lcd->setPixel(body_snake->x, body_snake->y, 1);
-            body_snake = body_snake->next;
-        }
-        lcd->setPixel(snake_food[0], snake_food[1], 1);
-        //
-        lcd->drawMap();
-    }
-    void updateBodyDirections()
-    {
-        snake.updateDirections();
-    }
-    void move()
-    {
-        checkCatchSeed();
-        updateBodyDirections();
-        //
-        snake.move();
-    }
-    void changeDirection(byte direction)
-    {
-        snake.setDirection(direction);
-    }
-    void growthSnake()
-    {
-        snake.growth();
-    }
-    void eatSeed()
-    {
-        growthSnake();
-    }
-    void checkCatchSeed()
-    {
-        BodySnake *body_snake = snake.getHead();
-        if( body_snake->x == snake_food[0] &&
-            body_snake->y == snake_food[1])
-        {
-            eatSeed();
-            generateSeed();
-        }
-    }
-    void generateSeed()
-    {
-        snake_food[0] = random(LCD_X);
-        snake_food[1] = random(LCD_Y);
-    }
-};
-
-const byte snakegame_load_image[6 * 84] = {
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0xC0, 0xC0, 0xE0, 0xE0, 0xE0, 0x80, 0x00, 0x00, 0x80, 0x80, 0x80, 0xC0, 0xC0, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0xC0, 0xC0, 0x00, 0x00, 0x00, 0x00, 0xC0, 0xC0, 0xC0, 0xE0, 0xC0, 0x00, 0x00, 0x80, 0xC0, 0xC0, 0xE0, 0xE0, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0xC0, 0xC0, 0x80, 0xC0, 0xC0, 0xC0, 0xE0, 0xE0, 0xE0, 0xC0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0xF8, 0xFC, 0xFF, 0xDF, 0xCF, 0x67, 0x67, 0x63, 0x63, 0xE3, 0xE1, 0xE0, 0xEC, 0x1F, 0xFF, 0xFF, 0x0F, 0x1F, 0x7F, 0xF8, 0xE0, 0xC0, 0xFF, 0xFF, 0xFF, 0x7F, 0xE0, 0xF8, 0xFC, 0xFF, 0x9F, 0xC7, 0xC3, 0xFF, 0xFF, 0xFC, 0xF8, 0xB8, 0x3B, 0x1F, 0xC7, 0xFF, 0xFF, 0xF8, 0xFC, 0xFE, 0xDF, 0x87, 0x03, 0x1F, 0x7D, 0xFF, 0xE3, 0xF3, 0x98, 0x90, 0x80, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x78, 0x3C, 0x3C, 0x1C, 0x1E, 0x1F, 0x0F, 0x0F, 0x07, 0x03, 0x0C, 0x87, 0x87, 0x83, 0x00, 0x80, 0x40, 0x40, 0xC3, 0x43, 0xC3, 0xC1, 0x45, 0xC4, 0xCF, 0xB7, 0x17, 0x93, 0xC0, 0x10, 0x50, 0x68, 0x09, 0x0F, 0x35, 0x45, 0x24, 0x0C, 0x4B, 0x4B, 0xCB, 0xD7, 0x10, 0x11, 0x33, 0xC3, 0x01, 0x00, 0x00, 0x03, 0x07, 0x03, 0x03, 0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x40, 0xC0, 0x60, 0xB0, 0xDC, 0xB4, 0xCA, 0xDA, 0x1E, 0xC9, 0xE9, 0x7F, 0xD4, 0x77, 0xFB, 0xF8, 0x43, 0x03, 0x82, 0x03, 0x82, 0x83, 0x05, 0x8D, 0x08, 0x58, 0x68, 0xB8, 0x08, 0x88, 0x88, 0x64, 0x30, 0x1C, 0x12, 0x8E, 0x86, 0x42, 0x22, 0x92, 0x90, 0x49, 0x2E, 0x08, 0x1E, 0x97, 0x08, 0xC6, 0xCA, 0x6A, 0xEC, 0xEC, 0xE8, 0x28, 0xE8, 0x68, 0x88, 0x10, 0x10, 0x60, 0xE0, 0xA0, 0xD0, 0x90, 0x90, 0x20, 0x40, 0x80, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x80, 0x87, 0xCF, 0xC8, 0xD7, 0xF7, 0xBC, 0xAF, 0x5D, 0x9C, 0xFF, 0x3C, 0x7C, 0xB3, 0xA0, 0xAF, 0xC3, 0x0C, 0xA7, 0x83, 0x30, 0xE3, 0xEC, 0xEA, 0x17, 0xC1, 0x38, 0x0C, 0x04, 0x02, 0xE1, 0xB1, 0x90, 0x6C, 0x74, 0xF2, 0xFA, 0xF9, 0xF8, 0xC0, 0xBF, 0xA1, 0x40, 0x40, 0x8A, 0x12, 0xB1, 0x4E, 0x59, 0xA7, 0x1A, 0x12, 0x93, 0x94, 0xA4, 0x27, 0x27, 0x27, 0x27, 0x24, 0xA6, 0xD4, 0x93, 0xF2, 0x1A, 0x9B, 0x75, 0x7F, 0xF8, 0xE0, 0x01, 0x02, 0xFE, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x07, 0x07, 0x07, 0x07, 0x0F, 0x0F, 0x0F, 0x0F, 0x0F, 0x0E, 0x0D, 0x09, 0x0B, 0x0A, 0x01, 0x0B, 0x0F, 0x0B, 0x0A, 0x05, 0x0E, 0x0F, 0x0F, 0x0F, 0x0C, 0x09, 0x06, 0x04, 0x0C, 0x08, 0x08, 0x08, 0x09, 0x11, 0x10, 0x10, 0x10, 0x12, 0x12, 0x13, 0x13, 0x13, 0x12, 0x12, 0x11, 0x11, 0x08, 0x08, 0x08, 0x05, 0x07, 0x0F, 0x0C, 0x0C, 0x05, 0x01, 0x01, 0x01, 0x01, 0x01, 0x05, 0x0C, 0x0F, 0x0C, 0x0B, 0x09, 0x09, 0x08, 0x0B, 0x0B, 0x08, 0x06, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
-};
